use modulo division to calculate ytics interval
i've used an if-else ladder, which did not anticipate values above 1
gigbabit and the script failed with an undefined variable in this case.
apparently, you don't need to set the ytics interval to the actual
interval you want to see on the axis but the interval modulo 10 is
enough and gnuplot will just omit any excess ytics? either way, there is
now no limit to the ymax value. (fixes #117)

if it turns out that this leads to graphical errors, i could use a
while-loop to get the interval into the correct magnitude:
  http://gnuplot.info/docs_5.5/loc18311.html